backend/sample_bills/VB-main/requirements.txt
# Core dependencies
PyYAML>=6.0
requests>=2.28.0
beautifulsoup4>=4.11.0

# Enhanced NLP capabilities
spacy>=3.4.0
nltk>=3.8.0
transformers>=4.20.0
torch>=1.12.0

# Database support (sqlite3 is built-in with Python)

# Advanced templating
Jinja2>=3.1.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Type checking and development
mypy>=1.0.0
black>=22.0.0
flake8>=5.0.0

# Performance and caching
cachetools>=5.0.0

# Web framework (for future API)
fastapi>=0.95.0
uvicorn>=0.20.0

# Additional utilities
python-dateutil>=2.8.0
colorama>=0.4.6  # For colored console output
tqdm>=4.64.0     # For progress bars
